+++ src/main/java/org/uom/lefterisxris/codetour/tours/ui/StepDescriptionEditor.java
@@ -7,10 +7,13 @@ import com.intellij.openapi.ui.Messages;
 import com.intellij.openapi.util.TextRange;
 import com.intellij.ui.components.JBScrollPane;
 import org.uom.lefterisxris.codetour.tours.domain.Step;
+import org.uom.lefterisxris.codetour.tours.service.StepRenderer;
 
 import javax.swing.*;
 import javax.swing.text.JTextComponent;
 import java.awt.*;
+import java.awt.event.MouseAdapter;
+import java.awt.event.MouseEvent;
 
 /**
  * Editor (as dialog) for Step's description value
@@ -20,10 +23,15 @@ import java.awt.*;
  */
 public class StepDescriptionEditor extends Messages.InputDialog {
 
+   private final Project project;
+   private final Step step;
+
    private StepDescriptionEditor(Project project, Step step) {
       super(project, String.format("Edit Step's '%s' description", step.getTitle()),
             "Edit Step", AllIcons.Actions.Edit, step.getDescription(), null,
             new String[]{Messages.getOkButton(), Messages.getCancelButton()}, 0);
+      this.project = project;
+      this.step = step;
    }
 
    public static String show(Project project, Step step, boolean preselected) {
@@ -60,6 +68,21 @@ public class StepDescriptionEditor extends Messages.InputDialog {
 
       myField = createTextFieldComponent();
       messagePanel.add(createScrollableTextComponent(), BorderLayout.CENTER);
+
+      final JButton preview = new JButton("Preview");
+      preview.addMouseListener(new MouseAdapter() {
+         @Override
+         public void mouseReleased(MouseEvent e) {
+            if (project != null && step != null)
+               StepRenderer.getInstance(Step.builder()
+                     .title(step.getTitle())
+                     .description(getTextField().getText())
+                     .file(step.getFile())
+                     .line(step.getLine())
+                     .build(), project, false).show();
+         }
+      });
+      messagePanel.add(preview, BorderLayout.SOUTH);
       return messagePanel;
    }